# openapi.yaml
openapi: 3.0.3
info:
  title: Lightning Engine API
  description: |
    REST API for the Lightning Engine ECS game server.

    ## Overview

    The engine uses a tick-based simulation model:
    1. **Commands** are queued via POST requests
    2. **Tick** advances process commands and run all module systems
    3. **Snapshots** return the current ECS state

    For real-time updates, connect via WebSocket at `ws://localhost:8080/snapshots/{matchId}`.

  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/matches/{matchId}:
    get:
      tags: [Matches]
      summary: Get match by ID
      operationId: getMatch
      parameters:
        - $ref: '#/components/parameters/matchId'
      responses:
        '200':
          description: Match details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResponse'
        '404':
          description: Match not found
    delete:
      tags: [Matches]
      summary: Delete a match
      description: Deletes match and all associated entities
      operationId: deleteMatch
      parameters:
        - $ref: '#/components/parameters/matchId'
      responses:
        '204':
          description: Match deleted
        '404':
          description: Match not found

  # Snapshot endpoints
...
  /api/snapshots/match/{matchId}:
    get:
      tags: [Snapshots]
      summary: Get snapshot for match
      description: Returns current ECS state filtered to a specific match
      operationId: getMatchSnapshot
      parameters:
        - $ref: '#/components/parameters/matchId'
      responses:
        '200':
          description: Match snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotResponse'
              example:
                matchId: 1
                tick: 42
                data:
                  MoveModule:
                    POSITION_X: [100.0, 200.0, 150.0]
                    POSITION_Y: [50.0, 75.0, 100.0]
                    VELOCITY_X: [1.0, -1.0, 0.0]
                  SpawnModule:
                    ENTITY_TYPE: [1.0, 2.0, 1.0]
                    OWNER_ID: [1.0, 1.0, 2.0]
        '404':
          description: Match not found

  # Module endpoints
...
  /api/resources:
    get:
      tags: [Resources]
      summary: List all resources
      operationId: listResources
      responses:
        '200':
          description: All uploaded resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceResponse'
    post:
      tags: [Resources]
      summary: Upload resource
      description: Upload a texture or other binary resource
      operationId: uploadResource
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                resourceName:
                  type: string
                  description: Optional name (defaults to filename)
                resourceType:
                  type: string
                  enum: [TEXTURE]
                  default: TEXTURE
              required:
                - file
      responses:
        '201':
          description: Resource uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
  /api/resources/{resourceId}:
    get:
      tags: [Resources]
      summary: Get resource metadata
      operationId: getResource
      parameters:
        - $ref: '#/components/parameters/resourceId'
      responses:
        '200':
          description: Resource metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
        '404':
          description: Resource not found
    delete:
      tags: [Resources]
      summary: Delete resource
      operationId: deleteResource
      parameters:
        - $ref: '#/components/parameters/resourceId'
      responses:
        '204':
          description: Resource deleted
        '404':
          description: Resource not found
  /api/resources/{resourceId}/data:
    get:
      tags: [Resources]
      summary: Download resource data
      operationId: downloadResource
      parameters:
        - $ref: '#/components/parameters/resourceId'
      responses:
        '200':
          description: Resource binary data
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Resource not found
  /api/resources/{resourceId}/stream:
    get:
      tags: [Resources]
      summary: Stream resource data
      description: Streams resource in chunks for large files
      operationId: streamResource
      parameters:
        - $ref: '#/components/parameters/resourceId'
      responses:
        '200':
          description: Streamed resource data
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Resource not found

  # GUI endpoints
...
components:
  parameters:
    matchId:
      name: matchId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    resourceId:
      name: resourceId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    MatchResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        enabledModuleNames:
          type: array
          items:
            type: string
    SnapshotResponse:
      type: object
      properties:
        matchId:
          type: integer
          format: int64
        tick:
          type: integer
          format: int64
        data:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: array
              items:
                type: number
                format: float
          description: |
            Columnar ECS data organized by module.
            Structure: { "ModuleName": { "COMPONENT_NAME": [value1, value2, ...] } }
            Each array contains one value per entity, in entity order.
...
    ResourceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
          enum: [TEXTURE]
        size:
          type: integer
          description: Size in bytes
...